+++ packages/react/impression-area/src/useImpressionRef.tsx
@@ -1,5 +1,4 @@
-import { usePreservedCallback } from '@toss/react';
-import { useRefEffect } from '@toss/use-ref-effect';
+import { usePreservedCallback, useRefEffect } from '@toss/react';
 import { noop } from '@toss/utils';
 import debounce from 'lodash.debounce';
 import { useMemo, useRef } from 'react';
