+++ packages/react/react/package.json
@@ -19,7 +19,6 @@
   "dependencies": {
     "@babel/runtime": "^7.14.8",
     "@toss/storage": "workspace:^1.0.0",
-    "@toss/use-ref-effect": "workspace:^1.0.0",
     "@toss/utils": "workspace:^1.0.0",
     "classnames": "^2.3.1",
     "lodash": "^4.17.20",
