+++ packages/react/react/src/hooks/useResizeObserver.ts
@@ -1,5 +1,5 @@
-import { useRefEffect } from '@toss/use-ref-effect';
 import { usePreservedCallback } from './usePreservedCallback';
+import { useRefEffect } from './useRefEffect';
 
 export type OnResize = (entry: ResizeObserverEntry) => void;
