--- packages/react/use-preserved-callback/.ts-prunerc.json
@@ -1,4 +1,0 @@
-{
-  "error": true,
-  "ignore": "^src\/index.ts"
-}
