--- packages/react/use-preserved-callback/.vscode/settings.json
@@ -1,17 +1,0 @@
-{
-  "eslint.nodePath": "../../../.yarn/sdks",
-  "prettier.prettierPath": "../../../.yarn/sdks/prettier/index.js",
-  "typescript.tsdk": "../../../.yarn/sdks/typescript/lib",
-  "typescript.enablePromptUseWorkspaceTsdk": true,
-  "search.exclude": {
-    "**/.yarn": true,
-    "**/.pnp.*": true,
-    "**/.next": true
-  },
-  "editor.formatOnSave": true,
-  "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true,
-    "source.organizeImport": true
-  },
-  "prettier.configPath": "../../../.prettierrc"
-}
