--- packages/react/use-preserved-callback/package.json
@@ -1,61 +1,0 @@
-{
-  "name": "@toss/use-preserved-callback",
-  "version": "1.0.0",
-  "main": "./src/index.ts",
-  "exports": {
-    ".": {
-      "require": "./src/index.ts",
-      "import": "./src/index.ts"
-    },
-    "./package.json": "./package.json"
-  },
-  "publishConfig": {
-    "access": "public",
-    "exports": {
-      ".": {
-        "require": "./dist/index.js",
-        "import": "./esm/index.mjs",
-        "types": "./dist/index.d.ts"
-      },
-      "./package.json": "./package.json"
-    },
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts",
-    "module": "./esm/index.js",
-    "import": "./esm/index.js"
-  },
-  "scripts": {
-    "prepack": "yarn build",
-    "build": "rm -rf dist esm && tsc --declaration --emitDeclarationOnly --declarationDir dist && rollup -c rollup.config.js",
-    "test": "echo \"no test specified.\"",
-    "typecheck": "tsc --noEmit",
-    "find-deadcode": "ts-prune -p ./tsconfig.prune.json"
-  },
-  "devDependencies": {
-    "@babel/core": "^7",
-    "@babel/runtime": "^7.13.10",
-    "@slash/rollup-config": "workspace:^",
-    "@types/jest": "^28.1.8",
-    "@types/node": "^14.14.41",
-    "@types/react": "^18.0.15",
-    "babel-jest": "^29.0.1",
-    "concurrently": "^4.1.0",
-    "jest": "^29.0.1",
-    "jest-environment-jsdom": "^29",
-    "react": "^18.2.0",
-    "rollup": "^2.77.0",
-    "ts-prune": "^0.9.2",
-    "typescript": "^4.2.3"
-  },
-  "peerDependencies": {
-    "@types/react": "*",
-    "react": "^16 || ^17 || ^18"
-  },
-  "peerDependenciesMeta": {
-    "@types/react": {
-      "optional": true
-    }
-  },
-  "sideEffects": false,
-  "gitHead": "bd6da79df8f3177ac1f13eac0edbf5b7549ea3d3"
-}
