--- packages/react/use-preserved-callback/src/index.ts
@@ -1,2 +1,0 @@
-/** @tossdocs-ignore */
-export * from './usePreservedCallback';
