--- packages/react/use-preserved-callback/tsconfig.prune.json
@@ -1,5 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["src/**/*"],
-  "exclude": ["**/*.spec.ts", "**/*.spec.tsx", "**/*.stories.tsx"]
-}
