--- packages/react/use-ref-effect/.ts-prunerc.json
@@ -1,4 +1,0 @@
-{
-  "error": true,
-  "ignore": "^src\/index.ts"
-}
