--- packages/react/use-ref-effect/babel.config.js
@@ -1,1 +1,0 @@
-module.exports = { extends: '../../../babel.config.js' };
