--- packages/react/use-ref-effect/jest.config.js
@@ -1,4 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-module.exports = require('@tossteam/jest').config({
-  rootDir: __dirname,
-});
