--- packages/react/use-ref-effect/jest.setup.js
@@ -1,2 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-require('@tossteam/jest').setup.jsdom();
