--- packages/react/use-ref-effect/package.json
@@ -1,65 +1,0 @@
-{
-  "name": "@toss/use-ref-effect",
-  "version": "1.0.0",
-  "main": "./src/index.ts",
-  "exports": {
-    ".": {
-      "require": "./src/index.ts",
-      "import": "./src/index.ts"
-    },
-    "./package.json": "./package.json"
-  },
-  "publishConfig": {
-    "access": "public",
-    "exports": {
-      ".": {
-        "require": "./dist/index.js",
-        "import": "./esm/index.mjs",
-        "types": "./dist/index.d.ts"
-      },
-      "./package.json": "./package.json"
-    },
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts",
-    "module": "./esm/index.js",
-    "import": "./esm/index.mjs"
-  },
-  "scripts": {
-    "prepack": "yarn build",
-    "build": "rm -rf dist esm && tsc --declaration --emitDeclarationOnly --declarationDir dist && rollup -c rollup.config.js",
-    "test": "jest",
-    "typecheck": "tsc --noEmit",
-    "find-deadcode": "ts-prune -p ./tsconfig.prune.json"
-  },
-  "devDependencies": {
-    "@babel/core": "^7",
-    "@babel/preset-env": "^7.16.4",
-    "@babel/preset-typescript": "^7.18.6",
-    "@babel/runtime": "^7.16.3",
-    "@slash/rollup-config": "workspace:^",
-    "@testing-library/react": "^13.3.0",
-    "@tossteam/jest": "^17.187.2",
-    "@types/babel__core": "^7",
-    "@types/babel__preset-env": "^7",
-    "@types/jest": "^28.1.8",
-    "@types/node": "^14.14.41",
-    "babel-jest": "^29.0.1",
-    "concurrently": "^4.1.0",
-    "jest": "^29.0.1",
-    "jest-environment-jsdom": "^29",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "rollup": "^2.77.0",
-    "ts-prune": "^0.9.2",
-    "typescript": "^4.2.3"
-  },
-  "peerDependencies": {
-    "react": "^16 || ^17 || ^18"
-  },
-  "sideEffects": false,
-  "dependencies": {
-    "@toss/use-preserved-callback": "workspace:^1.0.0",
-    "@types/react": "^17.0.3"
-  },
-  "gitHead": "bd6da79df8f3177ac1f13eac0edbf5b7549ea3d3"
-}
