--- packages/react/use-ref-effect/rollup.config.js
@@ -1,5 +1,0 @@
-const { generateRollupConfig } = require('@slash/rollup-config');
-
-module.exports = generateRollupConfig({
-  packageDir: __dirname,
-});
