--- packages/react/use-ref-effect/src/index.ts
@@ -1,2 +1,0 @@
-/** @tossdocs-ignore */
-export * from './useRefEffect';
