--- packages/react/use-ref-effect/tsconfig.esm.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../../tsconfig.esm.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./esm"
-  },
-  "include": ["src"]
-}
