--- packages/react/use-ref-effect/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../../tsconfig.build.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist"
-  },
-  "include": ["src"]
-}
